requirements.txt
absl-py==0.11.0
arrow==0.17.0
astunparse==1.6.3
binaryornot==0.4.4
cached-property==1.5.2
cachetools==4.1.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
et-xmlfile==1.0.1
future==0.18.2
gast==0.3.3
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
imutils==0.5.3
jdcal==1.4.1
Jinja2==2.11.2
jinja2-time==0.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.46
openpyxl==3.0.5
opt-einsum==3.3.0
pandas==1.1.4
Pillow==8.0.1
poyo==0.5.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
text-unidecode==1.3
tifffile==2020.11.18
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tqdm==4.52.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xlrd==1.2.0
zipp==3.4.0
